+++ src/main/java/org/ladysnake/elmendorf/Elmendorf.java
@@ -22,32 +22,5 @@
  */
 package org.ladysnake.elmendorf;
 
-import net.fabricmc.fabric.impl.gametest.FabricGameTestModInitializer;
-import net.minecraft.test.TestFunctions;
-
-import java.lang.reflect.Field;
-import java.util.Map;
-
 public final class Elmendorf {
-    /**
-     * Registers a class as a test container. This method should be called during test mod initialization.
-     * @param testClass a class containing test methods, which was not already registered through other means
-     * @param modId the id for the mod to which the class belongs
-     * @throws IllegalStateException if the class already got registered through this method or through the entrypoint
-     * @deprecated use QuiltGameTest#registerTests
-     */
-    @Deprecated
-    public static void registerTestClass(Class<?> testClass, String modId) {
-        try {
-            // Use reflection to avoid hard dependency on fabric-gametest-api
-            Field field = FabricGameTestModInitializer.class.getDeclaredField("GAME_TEST_IDS");
-            field.setAccessible(true);
-            @SuppressWarnings("unchecked") Map<Class<?>, String> testOwners = (Map<Class<?>, String>) field.get(null);
-            if (testOwners.containsKey(testClass)) throw new IllegalStateException(testClass + " got registered twice");
-            testOwners.put(testClass, modId);
-            TestFunctions.register(testClass);
-        } catch (IllegalAccessException | NoSuchFieldException e) {
-            throw new RuntimeException(e);
-        }
-    }
 }