+++ src/main/java/org/ladysnake/elmendorf/ThrowingRunnable.java
@@ -1,0 +1,5 @@
+package org.ladysnake.elmendorf;
+
+public interface ThrowingRunnable {
+    void run() throws Throwable;
+}